# api/openapi.yaml
openapi: 3.1.1
info:
  title: Тише — минимальный API v1
  version: 0.1.0
  description: 'Контракт локального прототипа. JSON, cookie-сессии, CSRF. Один пользователь
    — один дневник. Непубличные пользовательские маршруты требуют user и работают только с собственными
    данными; admin дает лишь отдельные административные разрешения. Все ошибки — Error. Общие правила
    и сценарии: README.md.'
servers:
- url: /api/v1
paths:
  /session:
    get:
      operationId: getSession
      summary: Текущая сессия и CSRF-токен
      responses:
        '200':
          description: Успешный ответ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Error'
      description: 'Доступен до входа: user=null и CSRF-токен анонимной сессии. При необходимости устанавливает
        cookie. Cache-Control: no-store. Блокированная аутентифицированная сессия: 403 account_blocked,
        а не user=null.'
      security: []
    post:
      operationId: login
      summary: Войти
      responses:
        '200':
          description: Успешный ответ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Error'
      description: CSRF из GET /session обязателен. Ротирует cookie и csrf_token; неверные данные 401
        invalid_credentials, блокировка 403 account_blocked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      security:
      - sessionCookie: []
        csrfToken: []
    delete:
      operationId: logout
      summary: Выйти
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
      description: Отзывает текущую сессию и удаляет cookie; допускается для анонимной/заблокированной
        сессии. После выхода повторить GET /session.
      security:
      - sessionCookie: []
        csrfToken: []
...
components:
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          description: Машинный код; неизвестный код показывается через message.
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      required:
      - code
      - message
      - fields
      examples:
      - code: sleep_overlap
        message: Сон пересекается с другой записью.
        fields:
          start: Укажите время после окончания предыдущего сна.
    User:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
            minLength: 1
          uniqueItems: true
        blocked:
          type: boolean
        timezone:
          type: string
          description: Часовой пояс IANA, например Europe/Moscow.
        default_schedule_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
      required:
      - id
      - version
      - name
      - email
      - roles
      - blocked
      - timezone
      - default_schedule_id
    Session:
      type: object
      additionalProperties: false
      properties:
        user:
          anyOf:
          - $ref: '#/components/schemas/User'
          - type: 'null'
        csrf_token:
          type: string
          minLength: 1
      required:
      - user
      - csrf_token
      examples:
      - user:
          id: 10000000-0000-4000-8000-000000000001
          version: 1
          name: Анна
          email: anna@example.com
          roles:
          - user
          blocked: false
          timezone: Europe/Moscow
          default_schedule_id: 20000000-0000-4000-8000-000000000001
        csrf_token: example-only-not-a-real-token
    Login:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
          writeOnly: true
      required:
      - email
      - password
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Анонимная или аутентифицированная серверная сессия. Cookie HttpOnly; Secure; SameSite=Lax;
        Path=/; без Domain. Анонимная сессия разрешена только для register/login/logout.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    Error:
      description: Единая ошибка. Статусы и коды перечислены в README.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoContent:
      description: Операция выполнена; тело отсутствует.